WIP: use Uint8Array instead of converting to hexString

+++ src/types.ts
@@ -10,7 +10,7 @@ export type Chain = {
 export type RailgunAddressLike = `0zk1${string}`;
 
 export type AddressData = {
-  masterPublicKey: bigint;
+  masterPublicKey: Uint8Array;
   viewingPublicKey: Uint8Array;
   chain?: Optional<Chain>;
   version?: Optional<number>;
